KataOS template support: honor release bit on reply capability
A capability attached to an RTReply may have the CAP_RELEASE flag or'd in
to signal the cap should be deleted after the reply is done. This gives us
"move" semantics where before one only had "copy" semantics. This needs
to be done here because on return constrol is returned to the CAmkES
glue code.

NB: this support is enabled based on CAP_RELEASE being defined (and only
    for MCS-enabled systems); it would be better if it were tied to the
    connector.

Change-Id: If17cf4f98e3cdd99c8dcc3ce372df544c0b20995